Add files via upload
Here are three datasets that I have acquired. The Grootbos_insects is a field trap experiment I did in 2021 with spiders and ants. The treecover_loss data set shows tree loss in India from 1994 to 2023 which would be pretty interesting to include. If we do include it then I will get treeloss for South Africa and another country too. The Iberian_bird_occurrences is for bird species in America in 2017 or 2019 (not sure how useful it is but could be a useful pre-covid dataset.